// PizzaOrderingSystem/src/app/shared/enum/pizza-image-types.enum.ts
export enum PizzaImageTypes {
    Cheese = 'Cheese',
    Veggie = 'Veggie',
    Pepperoni = 'Pepperoni',
    Meat = 'Meat',
    Margherita = 'Margherita',

    CheeseImage = 'https://image.shutterstock.com/image-photo/pizza-margherita-260nw-633097292.jpg',
    VeggieImage = 'https://www.archanaskitchen.com/images/archanaskitchen/1-Author/Waagmi_Soni/Gralic_Crust_Veggie_Pizza.jpg',
    PepperoniImage = 'https://media.istockphoto.com/photos/hot-homemade-pepperoni-pizza-picture-id521404215',
    MeatImage = 'https://iamafoodblog.b-cdn.net/wp-content/uploads/2015/10/meatlovers-pizza-8w.jpg',
    MargheritaImage = 'https://cook.fnr.sndimg.com/content/dam/images/cook/fullset/2011/1/6/0/CCEV103_Margherita-Pizza_s4x3.jpg.rend.hgtvcom.616.462.suffix/1416867304309.jpeg'
}